# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMRS RSUD Dr. H. Chasan Boesoirie Ternate - API Documentation",
...
        "contact": {
            "name": "IT SIMRS RSUD Dr. H. Chasan Boesoirie",
            "url": "https://chasanboesoirie.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/service/medifirst2000",
            "description": "Server SIMRS (Proxy Anti-CORS - Aktif Otomatis)"
        },
        {
            "url": "https://chasanboesoirie.id/service/medifirst2000",
            "description": "Server Langsung (chasanboesoirie.id - Butuh Postman/No-CORS)"
        }
    ],
    "paths": {
...
        "/kasir/daftar-tagihan-pasien": {
            "get": {
                "tags": [
                    "Tagihan & Billing Kasir Pasien"
                ],
                "summary": "Daftar Tagihan & Rekap Billing Pasien",
                "description": "Mengambil daftar tagihan pembayaran seluruh pasien yang teregistrasi. Menampilkan Nomor Registrasi, No CM, Nama Pasien, Tanggal Masuk, Tanggal Pulang, Ruangan, Kelas Perawatan, Penjamin (BPJS/Umum), Total Tagihan yang Harus Dibayar, dan Total Dijamin Rekanan/Asuransi.",
                "operationId": "72ac05ea26fadb5abdd091162f24584d",
                "parameters": [
                    {
                        "name": "noReg",
                        "in": "query",
                        "description": "Filter Nomor Registrasi Pasien",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2605000074"
                        }
                    },
                    {
                        "name": "noRm",
                        "in": "query",
                        "description": "Filter Nomor Rekam Medis (No CM)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "0487727"
                        }
                    },
                    {
                        "name": "tglAwal",
                        "in": "query",
                        "description": "Filter Tanggal Awal Struk / Masuk (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-01"
                        }
                    },
                    {
                        "name": "tglAkhir",
                        "in": "query",
                        "description": "Filter Tanggal Akhir Struk / Masuk (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-04"
                        }
                    },
                    {
                        "name": "instalasiId",
                        "in": "query",
                        "description": "Filter ID Instalasi / Departemen",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tagihan pasien berhasil dimuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "noregistrasi": {
                                                "type": "string",
                                                "example": "2605000074"
                                            },
                                            "nocm": {
                                                "type": "string",
                                                "example": "0487727"
                                            },
                                            "namapasien": {
                                                "type": "string",
                                                "example": "ADITHA BASRA NY"
                                            },
                                            "tglregistrasi": {
                                                "type": "string",
                                                "example": "2026-05-04 02:47:00"
                                            },
                                            "tglpulang": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "namaruangan": {
                                                "type": "string",
                                                "example": "IGD"
                                            },
                                            "namakelas": {
                                                "type": "string",
                                                "example": "Kelas III"
                                            },
                                            "kelompokpasien": {
                                                "type": "string",
                                                "example": "BPJS"
                                            },
                                            "norec": {
                                                "description": "No Rec Struk Pelayanan",
                                                "type": "string",
                                                "example": "sp-772910"
                                            },
                                            "nostruk": {
                                                "type": "string",
                                                "example": "STRUK/2605/0001"
                                            },
                                            "totalharusdibayar": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 350000
                                            },
                                            "totalprekanan": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 350000
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/kasir/detail-tagihan-pasien": {
            "get": {
                "tags": [
                    "Tagihan & Billing Kasir Pasien"
                ],
                "summary": "Rincian Tagihan Biaya Tindakan & Struk Pelayanan",
                "description": "Mengambil rincian per baris item tindakan, obat, dan layanan yang ditagihkan beserta harga, diskon, deposit, dan subtotal per struk pelayanan kasir.",
                "operationId": "d33688051d175ac3f639eb13a613da92",
                "parameters": [
                    {
                        "name": "noRecStrukPelayanan",
                        "in": "query",
                        "description": "No Rec Struk Pelayanan Pasien (sp.norec)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "sp-772910"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rincian tagihan tindakan pasien",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "noRegistrasi": {
                                            "type": "string",
                                            "example": "2605000074"
                                        },
                                        "noCm": {
                                            "type": "string",
                                            "example": "0487727"
                                        },
                                        "namaPasien": {
                                            "type": "string",
                                            "example": "ADITHA BASRA NY"
                                        },
                                        "jenisPenjamin": {
                                            "type": "string",
                                            "example": "BPJS"
                                        },
                                        "totalDeposit": {
                                            "type": "number",
                                            "example": 0
                                        },
                                        "jumlahBayar": {
                                            "type": "number",
                                            "example": 350000
                                        },
                                        "totalPenjamin": {
                                            "type": "number",
                                            "example": 350000
                                        },
                                        "detailTagihan": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "namaLayanan": {
                                                        "type": "string",
                                                        "example": "Pemeriksaan Dokter IGD"
                                                    },
                                                    "ruangan": {
                                                        "type": "string",
                                                        "example": "IGD"
                                                    },
                                                    "jumlah": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "harga": {
                                                        "type": "number",
                                                        "example": 150000
                                                    },
                                                    "diskon": {
                                                        "type": "number",
                                                        "example": 0
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "example": 150000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/kasir/daftar-pasien-aktif": {
            "get": {
                "tags": [
                    "Tagihan & Billing Kasir Pasien"
                ],
                "summary": "Daftar Pasien Aktif Dirawat (Billing Terbuka)",
                "description": "Mengambil daftar pasien yang masih aktif menjalani perawatan di rumah sakit dan belum menyelesaikan pembayaran billing kasir.",
                "operationId": "a05fc65d32649c77e3e244e1443bf2c0",
                "responses": {
                    "200": {
                        "description": "Daftar pasien aktif dirawat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Token otorisasi SIMRS Ternate. Default token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJoaXMuamtuIn0.FhnuCCWQx9SOMwBcLx3NFMPH45-KUwkblNKQY9Debm26PJ7ygt4-1z7oSMrefb0qfJwKtp02kS1O0lupFcMz1Q",
                "name": "X-AUTH-TOKEN",
                "in": "header"
            }
        }
    },
...
}